Add partial scan support
Notable changes:

* Workspace doesn't assume we include all files by default. Now we have
  to explicitly do that.
* Fixed a bug when removing all files in a directory it became fully
  included.
* Simplified inclusion/exclusion state tracking. Main features:

  * It's about as performant as the old one
  * It may consume a bit more memory on big code bases
  * It prevents project root escape
  * It may produce a more compact resulting path set

* Added a partial scan mode

Partial Scan mode
=================

This is a new way to construct a workspace for analysis. It's similar to
target file mode when one can specify files for analysis except it takes
into account `exclude_paths`.

This is useful in particular for editor integration where one would want
to get edited files automatically analyzed but not get reports for
excluded files. The results of a partial scan are always a subset of
a full scan.